# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tour Booking API
  version: 1.0.0
  description: RPC-over-HTTP API for seat-limited tour bookings with time-bound reservation holds, waitlists, and inventory adjustments
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.staging.example.com
    description: Staging environment
security:
  - BearerAuth: []
...
paths:
...
  /v1/waitlist/join:
    post:
      tags: [waitlist]
      summary: Join departure waitlist
      operationId: joinWaitlist
      x-idempotent: true
      x-idempotent-key: customer_ref_departure_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinWaitlistRequest'
            examples:
              example1:
                value:
                  departure_id: "dep_123"
                  customer_ref: "cust_EVG"
      responses:
        '200':
          description: Joined waitlist successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntry'
        '404':
          description: Departure not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/DefaultError'
  /v1/waitlist/notify:
    post:
      tags: [waitlist]
      summary: Process waitlist notifications (internal)
      operationId: notifyWaitlist
      x-idempotent: true
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotifyWaitlistRequest'
      responses:
        '200':
          description: Waitlist processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotifyWaitlistResponse'
        default:
          $ref: '#/components/responses/DefaultError'
  /v1/inventory/adjust:
    post:
      tags: [inventory]
      summary: Adjust departure capacity
      operationId: adjustInventory
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustInventoryRequest'
            examples:
              increase:
                value:
                  departure_id: "dep_123"
                  delta: 10
                  reason: "Upgraded to larger bus"
              decrease:
                value:
                  departure_id: "dep_123"
                  delta: -5
                  reason: "Bus maintenance reduced capacity"
      responses:
        '200':
          description: Inventory adjusted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryAdjustment'
        '409':
          description: Capacity conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                conflict:
                  value:
                    type: "https://docs.example.com/errors/conflict"
                    title: "Capacity conflict"
                    status: 409
                    code: "CAPACITY_CONFLICT"
                    detail: "Cannot reduce capacity below active holds"
                    trace_id: "c0ffee-9999"
        '422':
          description: Idempotency key mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/DefaultError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
      description: Unique key for idempotent operations
  schemas:
...
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          description: Problem type URI
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: URI reference for this occurrence
        code:
          type: string
          description: Application-specific error code
        retryable:
          type: boolean
          description: Whether the operation can be retried
        trace_id:
          type: string
          description: Trace ID for debugging
        violations:
          type: array
          description: Validation errors
          items:
            type: object
            properties:
              path:
                type: string
                description: JSON path to the invalid field
              message:
                type: string
                description: Validation error message
...
    Hold:
      type: object
      required: [id, departure_id, seats, customer_ref, status, expires_at]
      properties:
        id:
          type: string
          description: Unique hold ID
        departure_id:
          type: string
          description: Associated departure ID
        seats:
          type: integer
          minimum: 1
          description: Number of seats held
        customer_ref:
          type: string
          description: Customer reference
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CONFIRMED, CANCELED]
          description: Hold status
        expires_at:
          type: string
          format: date-time
          description: Hold expiration time (ISO 8601)
...
    WaitlistEntry:
      type: object
      required: [id, departure_id, customer_ref, created_at]
      properties:
        id:
          type: string
          description: Unique waitlist entry ID
        departure_id:
          type: string
          description: Associated departure ID
        customer_ref:
          type: string
          description: Customer reference
        created_at:
          type: string
          format: date-time
          description: Entry creation time (ISO 8601)
        notified_at:
          type: string
          format: date-time
          description: Notification time (ISO 8601)
    InventoryAdjustment:
      type: object
      required: [id, departure_id, delta, reason, created_at, actor]
      properties:
        id:
          type: string
          description: Unique adjustment ID
        departure_id:
          type: string
          description: Associated departure ID
        delta:
          type: integer
          description: Capacity change (positive or negative)
        reason:
          type: string
          description: Reason for adjustment
        created_at:
          type: string
          format: date-time
          description: Adjustment time (ISO 8601)
        actor:
          type: string
          description: User who made the adjustment
...
    JoinWaitlistRequest:
      type: object
      required: [departure_id, customer_ref]
      properties:
        departure_id:
          type: string
          description: Departure to join waitlist for
        customer_ref:
          type: string
          maxLength: 128
          description: Customer reference
    NotifyWaitlistRequest:
      type: object
      required: [departure_id]
      properties:
        departure_id:
          type: string
          description: Departure to process waitlist for
    NotifyWaitlistResponse:
      type: object
      required: [processed_count]
      properties:
        processed_count:
          type: integer
          description: Number of waitlist entries processed
        holds_created:
          type: array
          items:
            $ref: '#/components/schemas/Hold'
    AdjustInventoryRequest:
      type: object
      required: [departure_id, delta, reason]
      properties:
        departure_id:
          type: string
          description: Departure to adjust
        delta:
          type: integer
          description: Capacity change (positive or negative)
        reason:
          type: string
          minLength: 1
          maxLength: 500
          description: Reason for adjustment
...
  responses:
    DefaultError:
      description: Unexpected error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'